Extract performance trend service boundary
