+++ src/app/services/performance_workspace_service.py
@@ -1,20 +1,14 @@
 from __future__ import annotations
 
-import asyncio
 from collections.abc import Sequence
 from dataclasses import dataclass
-from datetime import date
 from typing import Any, TypeAlias, cast
 
 from app.config import settings
-from app.contracts.performance_attribution import (
-    AttributionSummaryView,
-    PerformanceAttributionTrendResponse,
-)
+from app.contracts.performance_attribution import AttributionSummaryView
 from app.contracts.performance_contribution import ContributionSummaryView
 from app.contracts.performance_evidence import PerformanceEvidenceView
 from app.contracts.performance_workspace import (
-    PerformanceHorizonComparisonResponse,
     PerformanceWorkspaceDetailsResponse,
     PerformanceWorkspaceResponse,
     PerformanceWorkspaceSummaryResponse,
@@ -27,7 +21,6 @@ from app.middleware.server_timing import server_timing_span
 from app.services.async_ttl_cache import AsyncTtlCache
 from app.services.performance_workspace_attribution import (
     parse_attribution_result,
-    parse_attribution_trend_results,
 )
 from app.services.performance_workspace_benchmarks import (
     fetch_benchmark_context,
@@ -35,18 +28,12 @@ from app.services.performance_workspace_benchmarks import (
 )
 from app.services.performance_workspace_capabilities import build_workspace_capabilities
 from app.services.performance_workspace_context import (
-    AttributionTrendRequestContext,
-    HorizonComparisonRequestContext,
     WorkspaceBenchmarkContext,
     WorkspaceOverviewState,
     WorkspaceReportWindow,
     WorkspaceRequestContext,
     WorkspaceRequestParameters,
-    assemble_attribution_trend_request_context,
-    assemble_horizon_comparison_request_context,
     assemble_workspace_request_context,
-    build_attribution_trend_dimension_context,
-    build_horizon_chart_frequency_context,
     build_workspace_dimension_context,
 )
 from app.services.performance_workspace_contribution import (
@@ -54,7 +41,6 @@ from app.services.performance_workspace_contribution import (
     parse_contribution_result,
 )
 from app.services.performance_workspace_controls import (
-    build_attribution_trend_windows,
     resolve_requested_window,
     resolve_workspace_summary_request,
 )
@@ -69,10 +55,6 @@ from app.services.performance_workspace_evidence import (
     fetch_evidence_view_state,
     fetch_performance_evidence_artifact,
     resolve_evidence_view_response,
-)
-from app.services.performance_workspace_horizon import (
-    fetch_workspace_horizon_dependencies,
-    parse_horizon_comparison_result,
 )
 from app.services.performance_workspace_projection import (
     project_portfolio_performance_snapshot,
@@ -93,6 +75,9 @@ from app.services.performance_workspace_summary import (
     ParsedWorkspaceSummary,
     parse_workspace_summary_result,
 )
+from app.services.performance_workspace_trend_service import (
+    PerformanceWorkspaceTrendServiceMixin,
+)
 from app.services.workbench_service import WorkbenchService
 from app.services.workspace_client_protocols import (
     PerformanceWorkspaceAnalyticsClient,
@@ -111,7 +96,7 @@ class WorkspaceDetailViews:
     attribution_detail_result: GatheredResult | None
 
 
-class PerformanceWorkspaceService:
+class PerformanceWorkspaceService(PerformanceWorkspaceTrendServiceMixin):
     def __init__(
         self,
         workbench_service: WorkbenchService,
@@ -272,355 +257,6 @@ class PerformanceWorkspaceService:
             include_detail_blocks=False,
         )
         return project_portfolio_performance_snapshot(workspace)
-
-    async def get_performance_horizon_comparison(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        period: str,
-        detail_basis: str,
-        benchmark_code: str | None,
-        chart_frequency: str,
-        explicit_start_date: str | None = None,
-        explicit_end_date: str | None = None,
-    ) -> PerformanceHorizonComparisonResponse:
-        context = await self._build_horizon_comparison_request_context(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            period=period,
-            chart_frequency=chart_frequency,
-            benchmark_code=benchmark_code,
-            explicit_start_date=explicit_start_date,
-            explicit_end_date=explicit_end_date,
-        )
-        workspace_summary_result = await self._fetch_horizon_comparison_dependencies(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            detail_basis=detail_basis,
-            context=context,
-        )
-        rows, resolved_benchmark_code = self._parse_horizon_comparison_rows(
-            detail_basis=detail_basis,
-            context=context,
-            workspace_summary_result=workspace_summary_result,
-        )
-        return self._build_horizon_comparison_response(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            detail_basis=detail_basis,
-            context=context,
-            benchmark_code=resolved_benchmark_code or context.benchmark_code or benchmark_code,
-            rows=rows,
-        )
-
-    async def _fetch_horizon_comparison_dependencies(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        detail_basis: str,
-        context: HorizonComparisonRequestContext,
-    ) -> GatheredResult:
-        async with server_timing_span("perf-horizon"):
-            return await fetch_workspace_horizon_dependencies(
-                analytics_client=self._analytics_client,
-                portfolio_id=portfolio_id,
-                correlation_id=correlation_id,
-                report_end_date=context.report_end_date,
-                report_start_date=context.report_start_date.isoformat()
-                if context.effective_period == "EXPLICIT"
-                else None,
-                period=context.effective_period,
-                detail_basis=detail_basis,
-                benchmark_code=context.benchmark_code,
-                portfolio_currency=context.overview.portfolio.base_currency,
-                chart_frequency=context.chart_frequency,
-            )
-
-    def _parse_horizon_comparison_rows(
-        self,
-        *,
-        detail_basis: str,
-        context: HorizonComparisonRequestContext,
-        workspace_summary_result: GatheredResult,
-    ) -> tuple[list[Any], str | None]:
-        return parse_horizon_comparison_result(
-            result=workspace_summary_result,
-            requested_period=context.effective_period,
-            requested_report_start_date=context.report_start_date.isoformat()
-            if context.effective_period == "EXPLICIT"
-            else None,
-            requested_report_end_date=context.report_end_date
-            if context.effective_period == "EXPLICIT"
-            else None,
-            detail_basis=detail_basis,
-            warnings=context.warnings,
-            partial_failures=context.partial_failures,
-        )
-
-    async def _build_horizon_comparison_request_context(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        period: str,
-        chart_frequency: str,
-        benchmark_code: str | None,
-        explicit_start_date: str | None,
-        explicit_end_date: str | None,
-    ) -> HorizonComparisonRequestContext:
-        overview_state = await self._load_workspace_overview_state(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-        )
-        report_window = await self._build_workspace_report_window(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            overview_state=overview_state,
-            period=period,
-            explicit_start_date=explicit_start_date,
-            explicit_end_date=explicit_end_date,
-        )
-        chart_frequency_context = build_horizon_chart_frequency_context(
-            chart_frequency=chart_frequency,
-            warnings=overview_state.warnings,
-        )
-        benchmark_context = await self._build_workspace_benchmark_context(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            report_end_date=report_window.report_end_date,
-            portfolio_currency=overview_state.overview.portfolio.base_currency,
-            benchmark_code=benchmark_code,
-            include_benchmark_catalog=True,
-        )
-        return assemble_horizon_comparison_request_context(
-            overview_state=overview_state,
-            report_window=report_window,
-            chart_frequency_context=chart_frequency_context,
-            benchmark_context=benchmark_context,
-        )
-
-    def _build_horizon_comparison_response(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        detail_basis: str,
-        context: HorizonComparisonRequestContext,
-        benchmark_code: str | None,
-        rows: list[Any],
-    ) -> PerformanceHorizonComparisonResponse:
-        benchmark_options = parse_benchmark_catalog_result(
-            result=context.benchmark_catalog_result,
-            assigned_benchmark_code=benchmark_code,
-            warnings=context.warnings,
-            partial_failures=context.partial_failures,
-        )
-        return PerformanceHorizonComparisonResponse(
-            correlation_id=correlation_id,
-            contract_version=context.overview.contract_version,
-            portfolio_id=portfolio_id,
-            as_of_date=context.overview.as_of_date,
-            period=context.effective_period,
-            report_start_date=context.report_start_date.isoformat(),
-            report_end_date=context.report_end_date,
-            reporting_currency=context.overview.portfolio.base_currency,
-            detail_basis=detail_basis,
-            chart_frequency=context.chart_frequency,
-            requested_chart_frequency_supported=context.requested_chart_frequency_supported,
-            benchmark_code=benchmark_code,
-            benchmark_options=benchmark_options,
-            rows=rows,
-            warnings=context.warnings,
-            partial_failures=context.partial_failures,
-        )
-
-    async def get_performance_attribution_trend(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        period: str,
-        chart_frequency: str,
-        attribution_dimension: str,
-        detail_basis: str,
-        benchmark_code: str | None,
-        explicit_start_date: str | None = None,
-        explicit_end_date: str | None = None,
-    ) -> PerformanceAttributionTrendResponse:
-        context = await self._build_attribution_trend_request_context(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            period=period,
-            chart_frequency=chart_frequency,
-            attribution_dimension=attribution_dimension,
-            benchmark_code=benchmark_code,
-            explicit_start_date=explicit_start_date,
-            explicit_end_date=explicit_end_date,
-        )
-        if context.benchmark_code is None:
-            context.warnings.append("ATTRIBUTION_TREND_UNAVAILABLE_NO_BENCHMARK")
-            return self._build_attribution_trend_response(
-                portfolio_id=portfolio_id,
-                correlation_id=correlation_id,
-                detail_basis=detail_basis,
-                context=context,
-                rows=[],
-            )
-
-        rows = await self._build_attribution_trend_rows(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            detail_basis=detail_basis,
-            context=context,
-        )
-        return self._build_attribution_trend_response(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            detail_basis=detail_basis,
-            context=context,
-            rows=rows,
-        )
-
-    async def _build_attribution_trend_rows(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        detail_basis: str,
-        context: AttributionTrendRequestContext,
-    ) -> Sequence[Any]:
-        window_pairs = self._build_attribution_trend_window_pairs(context)
-        attribution_results = await self._fetch_attribution_trend_results(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            detail_basis=detail_basis,
-            context=context,
-            window_pairs=window_pairs,
-        )
-        return parse_attribution_trend_results(
-            results=attribution_results,
-            window_pairs=window_pairs,
-            chart_frequency=context.chart_frequency,
-            requested_period="EXPLICIT",
-            warnings=context.warnings,
-            partial_failures=context.partial_failures,
-        )
-
-    async def _build_attribution_trend_request_context(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        period: str,
-        chart_frequency: str,
-        attribution_dimension: str,
-        benchmark_code: str | None,
-        explicit_start_date: str | None,
-        explicit_end_date: str | None,
-    ) -> AttributionTrendRequestContext:
-        overview_state = await self._load_workspace_overview_state(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-        )
-        report_window = await self._build_workspace_report_window(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            overview_state=overview_state,
-            period=period,
-            explicit_start_date=explicit_start_date,
-            explicit_end_date=explicit_end_date,
-        )
-        dimension_context = build_attribution_trend_dimension_context(
-            chart_frequency=chart_frequency,
-            attribution_dimension=attribution_dimension,
-            warnings=overview_state.warnings,
-        )
-        benchmark_context = await self._build_workspace_benchmark_context(
-            portfolio_id=portfolio_id,
-            correlation_id=correlation_id,
-            report_end_date=report_window.report_end_date,
-            portfolio_currency=overview_state.overview.portfolio.base_currency,
-            benchmark_code=benchmark_code,
-            include_benchmark_catalog=False,
-        )
-        return assemble_attribution_trend_request_context(
-            overview_state=overview_state,
-            report_window=report_window,
-            dimension_context=dimension_context,
-            benchmark_context=benchmark_context,
-        )
-
-    def _build_attribution_trend_window_pairs(
-        self,
-        context: AttributionTrendRequestContext,
-    ) -> list[tuple[date, date]]:
-        return build_attribution_trend_windows(
-            start_date=context.report_start_date,
-            end_date=date.fromisoformat(context.report_end_date),
-            chart_frequency=context.chart_frequency,
-        )
-
-    async def _fetch_attribution_trend_results(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        detail_basis: str,
-        context: AttributionTrendRequestContext,
-        window_pairs: list[tuple[date, date]],
-    ) -> Sequence[GatheredResult]:
-        if context.benchmark_code is None:
-            return []
-        async with server_timing_span("perf-attribution"):
-            gathered = await asyncio.gather(
-                *(
-                    self._analytics_client.get_attribution_analytics(
-                        portfolio_id=portfolio_id,
-                        report_start_date=window_start.isoformat(),
-                        report_end_date=window_end.isoformat(),
-                        period="EXPLICIT",
-                        metric_basis=detail_basis,
-                        benchmark_id=context.benchmark_code,
-                        dimension=context.attribution_dimension,
-                        correlation_id=correlation_id,
-                    )
-                    for window_start, window_end in window_pairs
-                ),
-                return_exceptions=True,
-            )
-        return cast(Sequence[GatheredResult], gathered)
-
-    def _build_attribution_trend_response(
-        self,
-        *,
-        portfolio_id: str,
-        correlation_id: str,
-        detail_basis: str,
-        context: AttributionTrendRequestContext,
-        rows: Sequence[Any],
-    ) -> PerformanceAttributionTrendResponse:
-        return PerformanceAttributionTrendResponse(
-            correlation_id=correlation_id,
-            contract_version=context.overview.contract_version,
-            portfolio_id=portfolio_id,
-            as_of_date=context.overview.as_of_date,
-            period=context.effective_period,
-            report_start_date=context.report_start_date.isoformat(),
-            report_end_date=context.report_end_date,
-            chart_frequency=context.chart_frequency,
-            detail_basis=detail_basis,
-            attribution_dimension=context.attribution_dimension,
-            requested_chart_frequency_supported=context.requested_chart_frequency_supported,
-            requested_attribution_dimension_supported=(
-                context.requested_attribution_dimension_supported
-            ),
-            benchmark_code=context.benchmark_code,
-            rows=list(rows),
-            warnings=context.warnings,
-            partial_failures=context.partial_failures,
-        )
 
     async def _build_performance_workspace_response(
         self,
